More discrete source links (#22)
* Link source as an emoji next to the var name

Previously, the link to the source of a var was in a `<sub></sub>` at
the end of the var description. However, it did not always play nicely
with the Github markdown renderer. For instance, when the last thing in
a docstring was a markdown table, that link would appear as a row.

This commit solves that kind of issues by moving the link to an emoji
right next to the var name. It also declutters the rendered page and
make var names more striking.

* Print separators between namespaces

Making navigation easier when there are several namespaces on the
rendered page.